Varios

+++ 3-tecnologia/1-sistemas/2-STP-Stack-de-Planificación/1-SGA-sistema-de-gestion-de-activos/2-fuentes/abm/scripts/activador.sh
@@ -6,7 +6,6 @@ for arch in `cat activador.txt`;do
 	if [ -a ${arch} ]; then
 		sed 's/SetUseImagesForActions(false)/SetUseImagesForActions(true)/g' ${arch} | \
 			sed 's/SetHighlightRowAtHover(false)/SetHighlightRowAtHover(true)/g' | \
-			sed 's/SetShowPageList(false)/SetShowPageList(true)/g' | \
 			sed 's/SetExportToExcelAvailable(false)/SetExportToExcelAvailable(true)/g' | \
 			sed 's/SetExportToWordAvailable(false)/SetExportToWordAvailable(true)/g' | \
 			sed 's/SetExportToXmlAvailable(false)/SetExportToXmlAvailable(true)/g' | \
@@ -23,13 +22,18 @@ for arch in `cat activador.txt`;do
 			sed 's/AddBand(/AddBandToBegin(/g' | \
 			sed 's/SetRowsPerPage(.*)/SetRowsPerPage(100)/g' > ${arch}.new
 			
-		grep "?>" ${arch} >/dev/null 2>&1
-		if [ $? -ne 0 ]; then
-			cat ${arch}.new general.php ${arch%%.*}.cm? >${arch} 2>/dev/null
-			rm ${arch}.new
-		else
-			echo "-> archivo ya completado"
+		if [[ ${arch} == ver_* ]]; then
+			echo "-> No se completa el archivo ${arch}"
 			mv ${arch}.new ${arch}
+		else
+			grep "?>" ${arch} >/dev/null 2>&1
+			if [ $? -ne 0 ]; then
+				cat ${arch}.new general.php ${arch%%.*}.cm? >${arch} 2>/dev/null
+				rm ${arch}.new
+			else
+				echo "-> archivo ya completado"
+				mv ${arch}.new ${arch}
+			fi
 		fi
 	fi
 done
